pyproject.toml
[project]
name = "saas-backend"
version = "0.1.0"
description = "SaaS backend for portfolio and client workspace"
authors = [{name="Auto Entrepreneur"}]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "firebase-admin>=6.5.0",
    "httpx>=0.27.0",
    "ruff>=0.4.8",
    "fpdf2>=2.7.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "pytest-asyncio>=0.23.7", "mypy>=1.10.0"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
